Address review feedback: simplify check_weights/check_init, add backend tests
- check_weights: use dtype=float for default weights, let xp.asarray
  infer dtype for user-provided weights (avoids cross-backend dtype issues)
- check_init: preserve init.dtype when available
- nanmean_riemann: convert torch init to numpy before check_init
- Fix test_ajdpham: use real dtype for weights with HPD matrices
- Add tests for joint_eigvalsh and pairwise_euclidean in test_utils_backend

+++ tests/test_utils_ajd.py
@@ -74,7 +74,7 @@ def test_ajdpham_weight_none_equivalent_uniform(kind, get_mats):
     n_matrices, n_channels = 5, 3
     X = get_mats(n_matrices, n_channels, kind)
     xp = get_namespace(X)
-    ones = xp.ones(n_matrices, dtype=X.dtype, device=xpd(X))
+    ones = xp.ones(n_matrices, dtype=X.real.dtype, device=xpd(X))
     V, D = ajd_pham(X)
     Vw, Dw = ajd_pham(X, sample_weight=ones)
     assert_array_equal(V, Vw)
